requirements.txt
altgraph==0.17.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
future==1.0.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
macholib==1.16.3
MarkupSafe==3.0.2
packaging==24.2
requests==2.32.3
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
numpy>=1.24
Pillow>=9.5
mpmath=1.3.0
sympy=1.14.0